datathon_linechatbot/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
PyMySQL==0.10.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
